# .eh_frame: Don't allocate the full unwind table by default