As we are going to enable DWARF-based unwinding by default soon, which
means that the DWARF unwind information parser and the table generation
will be called significantly more times.

During prototyping, some shortcuts were taken, such as generating the
unwind table ahead of time. The unwind table is an array of unwind rows,
which for "medium-sized" binaries it can reach 200k entries. For the largest
binaries we've seen, the unwind tables were up to 2M entries big.

Generating and keeping these large slices in memory is not good for
performance, as we need to re-allocate them multiple times when
regrowing the backing memory buffer, as well as copying the data over, if
necessary.

While we could reasonably estimate the final size of the buffer using the
`.text` size as a proxy, this is not necessary for the majority of
cases, and we can process one unwind entry at a time.

This is what this commits implements. It refactors the table generation
code to change from returning said unwind table in a slice, to accepting
a callback function that can then process each element in the way it
finds it suitable.

Note that there are way more optimisations that we could do to speed up
the code but we'll first profile this code in production once we carry
out the monolithic unwind table work, as its architecture will reduce
the redundant work.

__Note__: The unwind generation table is still generating the full table ahead
of time. This would have to be changed to take advantage of this new API.

## Benchmarks

```
$ go test -v github.com/parca-dev/parca-agent/pkg/stack/unwind -bench=. -count=30
```

```
[javierhonduco@fedora parca-agent]$ benchstat /tmp/old /tmp/new
name                                  old time/op    new time/op    delta
ParsingLibcDwarfUnwindInformation-12    34.5ms ±43%    10.8ms ±32%  -68.60%  (p=0.000 n=30+30)

name                                  old alloc/op   new alloc/op   delta
ParsingLibcDwarfUnwindInformation-12    32.7MB ± 0%     5.6MB ± 0%  -82.80%  (p=0.000 n=30+29)

name                                  old allocs/op  new allocs/op  delta
ParsingLibcDwarfUnwindInformation-12      150k ± 0%       83k ± 0%  -44.76%  (p=0.000 n=30+30)
```