@javierhonduco 2022-12-08
As discussed in the chat, removing the callbacks resulted in a big perf regression (see commit msg) 😮. I couldn't spot anything obvious by inspecting the code.