@kakkoyun 2022-12-09
> As discussed in the chat, removing the callbacks resulted in a big perf regression (see commit msg) 😮. I couldn't spot anything obvious by inspecting the code.

Probably because of the closure we return. Have you tried to return the pointer directly?