@javierhonduco 2022-12-09
> Probably because of the closure we return. Have you tried to return the pointer directly?

Which one?