@kakkoyun 2022-12-09
> > Probably because of the closure we return. Have you tried to return the pointer directly?
> 
> Which one?

https://github.com/parca-dev/parca-agent/pull/1101/files#r1044656024