@javierhonduco 2022-12-12
The new approach that doesn't use callbacks has pretty much the same perf characteristics as the initial commit:

```
[javierhonduco@fedora parca-agent]$ benchstat /tmp/old /tmp/new2
name                                  old time/op    new time/op    delta
ParsingLibcDwarfUnwindInformation-12    58.9ms ±37%    13.4ms ±25%  -77.21%  (p=0.000 n=18+30)

name                                  old alloc/op   new alloc/op   delta
ParsingLibcDwarfUnwindInformation-12    32.7MB ± 0%     5.6MB ± 0%  -82.89%  (p=0.000 n=19+29)

name                                  old allocs/op  new allocs/op  delta
ParsingLibcDwarfUnwindInformation-12      150k ± 0%       79k ± 0%  -47.21%  (p=0.000 n=18+30)
```

Slightly better allocs/op 79k vs 83k, slightly worse timing but falls within the error margin 😄