@javierhonduco 2022-12-13
Thanks for the comprehensive survey on Go iterators! I don't particularly prefer any of them, so thinking of landing the current one and once Go's iterator proposal materializes we can use it